Rename python file to match standard casing
